Make alarms stream return the Alarms object
