+++ Cargo.toml
@@ -14,7 +14,7 @@ async-stream = "0.3"
 axum = { version = "0.8", features = ["ws"] }
 axum-server = { version = "0.7", features = ["tls-rustls"] }
 base64 = "0.22"
-chrono = "0.4.44"
+chrono = { version = "0.4.44", features = ["serde"] }
 futures = "0.3.27"
 futures-util = { version = "0.3", features = ["tokio-io"] }
 http = "1"
