+++ src/graphql/alarms/mod.rs
@@ -5,12 +5,12 @@
 use crate::{
     g_rpc::{alarms_db, alarms_svc, proto::services::alarms},
     graphql::alarms::types::Alarm,
-    pubsub::{Message, Subscriber, kafka_impl::KafkaSubscriber},
+    pubsub::{Subscriber, kafka_impl::KafkaSubscriber},
 };
 use async_graphql::{Error, Object, Subscription};
 use chrono::{DateTime, Utc};
 use rust_env_var_lib::env_var;
-use tokio_stream::wrappers::BroadcastStream;
+use tokio_stream::{Stream, StreamExt};
 use tonic::{Code, Request, Status};
 use tracing::error;
 use types::{AlarmGroup, AlarmGroupMetadatum, AlarmTimer, UserLayout};
@@ -221,9 +221,20 @@ pub struct AlarmsSubscriptions;
 #[Subscription]
 impl AlarmsSubscriptions {
     /// Streams back all alarms from the alarms topic.
-    async fn alarms(&self) -> Result<BroadcastStream<Message>, Error> {
+    async fn alarms(&self) -> Result<impl Stream<Item = Alarm>, Error> {
         KafkaSubscriber::subscribe(get_host(), get_topic())
             .await
+            .map(|stream| {
+                stream.filter_map(|stream_item| match stream_item {
+                    Err(e) => {
+                        error!("{e:?}");
+                        None
+                    }
+                    Ok(message) => Alarm::try_from(message)
+                        .inspect_err(|e| error!("{e:?}"))
+                        .ok(),
+                })
+            })
             .map_err(|e| Error::new(format!("{e}")))
     }
 }
